Explain the absence of a vmimport_role_name variable
When you have a role description variable you should usually have a role name
variable as well. However, in this case we need to use a static value because
the nature of the `vmimport` service role. AWS requires a specific name for the
role to be usable by the VM Import/Export feature so it should not be changed.

Co-authored-by: dav3r <david.redmin@trio.dhs.gov>